+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/ifaceresource/support/InterfaceResourceApp.groovy
@@ -1,0 +1,26 @@
+package ru.vyarus.dropwizard.guice.cases.ifaceresource.support
+
+import io.dropwizard.Application
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.support.TestConfiguration
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 18.06.2016
+ */
+class InterfaceResourceApp extends Application<TestConfiguration> {
+
+    @Override
+    void initialize(Bootstrap<TestConfiguration> bootstrap) {
+        bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
+                .enableAutoConfig(this.class.package.name)
+                .build()
+        );
+    }
+
+    @Override
+    void run(TestConfiguration configuration, Environment environment) throws Exception {
+    }
+}