+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/ifaceresource/support/InvisibleResourceImpl.groovy
@@ -1,0 +1,15 @@
+package ru.vyarus.dropwizard.guice.cases.ifaceresource.support
+
+/**
+ * This resource could not be recognized because @Path annotation is not on 1st level interface.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 18.06.2016
+ */
+class InvisibleResourceImpl implements SecondLevelResource {
+
+    @Override
+    String latest() {
+        return null
+    }
+}