+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/ifaceresource/support/ResourceImpl.groovy
@@ -1,0 +1,12 @@
+package ru.vyarus.dropwizard.guice.cases.ifaceresource.support
+/**
+ * @author Vyacheslav Rusakov
+ * @since 18.06.2016
+ */
+class ResourceImpl implements ResourceContract {
+
+    @Override
+    String latest() {
+        return "called!"
+    }
+}