+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/ifaceresource/support/SecondLevelResource.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.dropwizard.guice.cases.ifaceresource.support
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 18.06.2016
+ */
+interface SecondLevelResource extends ResourceContract {
+
+}