+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Update to guice 4.1.0
+* ResourceInstaller looks for @Path on directly implemented interface (#10)
 
 ### 3.2.0 (2016-01-23)
 * Clear possible duplicate guicey bundle instances